+++ mindspore/lite/tools/converter/parser/tflite/tflite_addn_parser.cc
@@ -49,7 +49,7 @@ STATUS TfliteAddNParser::Parse(const std::unique_ptr<tflite::OperatorT> &tflite_
   op->primitive->value.value = attr.release();
 
   // set input
-  for (int i = 0; i < tflite_op->inputs.size(); i++) {
+  for (size_t i = 0; i < tflite_op->inputs.size(); i++) {
     AddOpInput(op, tensors_id, tensors_format, tensors_id_map,
                tflite_op->inputs[i], tensors_id->size(), tflite_tensors.size(), schema::Format_NHWC);
   }
